requirements.txt
sqlalchemy==2.0.30
alembic==1.16.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1